# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "some api",
        "version": "0.1"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
        "/catalog/products": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Returns list of products",
                        "schema": {
                            "properties": {
                                "next": {
                                    "description": "reference to next page of results",
                                    "type": "string"
                                },
                                "prev": {
                                    "description": "reference to previous page of results",
                                    "type": "string"
                                },
                                "data": {
                                    "description": "products list",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/ProductListItem"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NewProduct"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Product created",
                        "schema": {
                            "type": "array",
                            "items": {}
                        }
                    }
                }
            }
        },
        "/catalog/products/{productId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "162e2dc2-6761-4a4e-9203-05f367d7ccd9"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns product",
                        "schema": {
                            "$ref": "#/definitions/ProductListItem"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "162e2dc2-6761-4a4e-9203-05f367d7ccd9"
                        ]
                    },
                    {
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateProduct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product updated",
                        "schema": {
                            "type": "array",
                            "items": {}
                        }
                    }
                }
            }
        },
        "/checkout/carts": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "Cart created",
                        "schema": {
                            "type": "array",
                            "items": {}
                        }
                    }
                }
            }
        },
        "/checkout/carts/{cartId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "cartId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "3d73fbef-7998-4836-a521-004fdfbb0241"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns single cart with products",
                        "schema": {
                            "$ref": "#/definitions/Cart"
                        }
                    }
                }
            }
        },
        "/checkout/carts/{cartId}/products": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "cartId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "3d73fbef-7998-4836-a521-004fdfbb0241"
                        ]
                    },
                    {
                        "name": "addProductToCart",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AddProductToCart"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Product added to cart",
                        "schema": {
                            "type": "array",
                            "items": {}
                        }
                    }
                }
            }
        },
        "/checkout/carts/{cartId}/products/{productId}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "cartId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "3d73fbef-7998-4836-a521-004fdfbb0241"
                        ]
                    },
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "162e2dc2-6761-4a4e-9203-05f367d7ccd9"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product removed from cart",
                        "schema": {
                            "type": "array",
                            "items": {}
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ProductListItem": {
            "required": [
                "id",
                "name",
                "price"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "66e3a13c-d1a7-4bc0-9732-b99c184602e7"
                },
                "name": {
                    "type": "string",
                    "example": "foo"
                },
                "price": {
                    "type": "number",
                    "example": "9.99"
                }
            },
            "type": "object"
        },
        "NewProduct": {
            "required": [
                "name",
                "price"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "minLength": 1,
                    "example": "foo"
                },
                "price": {
                    "type": "number",
                    "format": "decimal",
                    "minimum": 0.01,
                    "example": "9.99"
                }
            },
            "type": "object"
        },
        "UpdateProduct": {
            "properties": {
                "name": {
                    "type": "string",
                    "minLength": 1,
                    "example": "foo"
                },
                "price": {
                    "type": "number",
                    "format": "decimal",
                    "minimum": 0.01,
                    "example": "9.99"
                }
            },
            "type": "object"
        },
        "Cart": {
            "required": [
                "id",
                "products",
                "total"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "0ecdc635-bb14-4ffa-8826-756d9cc3c73d"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ProductCartItem"
                    }
                },
                "total": {
                    "type": "number",
                    "minimum": 0,
                    "example": "9.99"
                }
            },
            "type": "object"
        },
        "ProductCartItem": {
            "required": [
                "id",
                "name",
                "price"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "66e3a13c-d1a7-4bc0-9732-b99c184602e7"
                },
                "name": {
                    "type": "string",
                    "example": "foo"
                },
                "price": {
                    "type": "number",
                    "minimum": 0.01,
                    "example": "9.99"
                }
            },
            "type": "object"
        },
        "AddProductToCart": {
            "required": [
                "product",
                "quantity"
            ],
            "properties": {
                "product": {
                    "type": "string",
                    "example": "7dbaf7f6-c415-42cf-85c2-9a8fababcba6"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                }
            },
            "type": "object"
        }
    }
}